refactor: move getTopLevelVariant
into inventory plugin
